fix: manually summarise datapoints for graphiteWorking check
Using graphite's `summarize` function gives you 2 datapoints, one for the left-most part of the axis
and the other for the right-most. These are averaged out to give a straight line on a graph, but don't
represent accurate values, for example if the latest value is 0 but the rest are high, you won't see
the dropoff using `summarize` until the high data is over the bounds of the average. Conversely, if its
at 0 for a while, and new non-zeo data comes in, again this won't be visible or accurate for a while.

The purpose of this healthcheck is to simply check if the latest entries are non-zero. So we can potentially
just get the latest value in the timeseries and check that... a small problem herein lies that the data
in graphite can be out of sync from real world. From eye balling as much as 7 minute delays before it
starts backfilling data.

Because of this, the function has now been changed to get the last 10 minutes worth of data, sum this up
in JS and getting the grand total number of requests over the last 10 minutes. If _that_ number is 0 then
we have a problem. This means we'll know about failures 10 minutes after they happen. 10 minutes seems
like a long time but it is a trade off between long times knowing about failure vs lots of erroneous false
positives (or negatives?).