Use latest datapoint instead of summing
Currently if multiple datapoints are returned, they are summed and that value
is used to compare with the threshold. This doesn't make much sense for many
CloudWatch statistics and could be quite misleading if your intention was
to compare the threshold with the metric over the period that you've specified.

Instead of doing this, use the latest datapoint available, and compare that
metric against the threshold.