Use a larger time window for 60 second period
CloudWatch is a bit slack and sometimes takes a few seconds to populate
new datapoints. This can mean that you sometimes get an empty result set
if you query CW at the start of a minute with a 60s period and your start
time is 60 seconds ago.

This commit extends the window to 90 seconds in this situation,
so you should always get at least one datapoint in the response.