Add an option to ignoreNullData to graphiteThreshold
Recently on ft.com we had an incident in which lots of slices
disappeared from the homepage.

Ordinarily we have monitoring in the next-front-page application
which would have caught this - but the data in Graphite is null
when the slices disappear which meant that the health check
never failed, so we relied on editorial to inform us of this
incident.

In my opinion missing/null data for a healthcheck is more often
a sign of a problem than not. However n-health is explicit about
ignoring null data, so I thought it was best to add an option
rather than reversing the behaviour entirely.