Log null data in GraphiteThreshold check
In dev huddle today I spoke about the fact that historically we don't
alert based on null data.

Some people mentioned this may be due to the reliability of Graphite
in the past. I'd like to log the rate that we encounter null data
as part of this threshold check with the potential to reverse the
behaviour if we find that it's infrequent.