Require n-logger v10.2.0
n-logger v10.2.0 adds in the ability to switch to Heroku log drains
rather than sending logs to Splunk manually. In order to support this
feature we need to limit the version range to be v10.2.0 or above
otherwise the feature will be present inconsistently.

This is not a breaking change because the major versions of n-logger
between 6 and 10 have no impact on this package:

  - v7.0.0 of n-logger only supports Node.js 14+ (no change for us)
  - v8.0.0 of n-logger re-adds support for Node.js 12
  - v9.0.0 of n-logger deprecates the SYSTEM_CODE environment variable
    (but then undeprecates it in v9.0.1)
  - v10.0.0 of n-logger only supports Node.js 14+ (no change for us)

We also needed to bump n-raven to a version which includes the latest
version of n-logger, to ensure that any logs in the dependencies are
formatted correctly for the log drain.

This is also not a breaking change because the major versions of
n-raven between 2 and 6 have no impact on the public API of this
package:

  - v3.0.0 of n-raven makes a lot of internal code changes (with no good
    change documentation), however the public parts of the Raven API
    used by n-health are still in place with the same method signatures.
    I don't think there's a breaking change here.
  - v4.0.0 of n-raven only supports Node.js 14+ (no change for us)
  - v5.0.0 of n-raven re-adds support for Node.js 12
  - v6.0.0 of n-raven only supports Node.js 14+ (no change for us)