Add a Heroku log drain check type
This adds a new check type which can validate that a Heroku app has a
valid log drain configured. We're validating against the configuration
[documented here](https://tech.in.ft.com/tech-topics/logging/splunk/logging-from-heroku#configuration).

One potentially contentious decision is use of the `HEROKU_APP_ID`
environment variable. Not all of our apps have this, and we'll need the
value in Heroku to be set to the _specific_ app so different per region.
I believe the ones that have this value are using this Heroku beta?
https://devcenter.heroku.com/articles/dyno-metadata

We have two options to resolve this:

1. We enable this beta for all our apps

2. We attempt to _guess_ the app name in n-express based on the `REGION`
   and `SYSTEM_CODE` environment variables, but this relies on apps
   using our established conventions consistently.