Remove n-raven
The use of n-raven in this library is causing us some issues rolling out
log drains. As part of the move to log drains we want to consistently
log errors which crashed the application and we're having to do this
work in n-raven.

While n-raven is included in both n-health and n-express, it's easy for
the versions to get out of sync. If this happens then the Sentry Raven
client is installed twice, which (because we're relying on a singleton)
will actually remove any custom unhandled exception handling code and
revert to the default Sentry handler with no error logging.

This is really bad and it could happen quite easily without anyone
knowing.

We decided that the best approach would be to remove Sentry logging from
n-health entirely and rely only on the errors we log to Splunk. A later
n-health PR will introduce Reliability Kit's error logging which will
improve things further.

This PR will result in a new major version of n-health, because we think
logging to Sentry forms part of the public API and it should be a
concious decision for teams to remove it.

Co-Authored-By: Alex Muller <alex.muller@ft.com>
Co-Authored-By: Andy Gout <andy.gout@ft.com>