Remove the graphite checks
This is a breaking change. We're removing the various health check types
that rely on reading Graphite metrics because, as of Q3 2024, we will no
longer be able to read this data.

For now we recommend removing these checks from your apps. The alerts
aren't very commonly-used and we have better ways of monitoring uptime
now. If you still need alerts based on Graphite metrics then you can do
so by creating a dashboard in the grafana monorepo and adding alerts:

https://github.com/Financial-Times/grafana/tree/main/terraform/dashboards

See-Also: https://financialtimes.atlassian.net/browse/CPREL-1056