Remove Chai
Chai v5 is ESM-only which would be a really big migration. We're not
doing any complicated assertions in the tests here so I think it makes
sense to just switch to the built-in `assert` module. We get to drop a
dependency (just under 1MB) and I think assert is easier to understand
anyway.